+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Unreleased
 
-* Add support for dynamic instanced filtered graphs using filter instance ids encoded in channel descriptors.
+* Add support for dynamic/static instanced filtered graphs using filter instance ids encoded in channel descriptors.
 
 ### [v6.164](https://github.com/replicant4j/replicant/tree/v6.164) (2026-01-08) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.163...v6.164)